requirements.txt
# VK
vk-api>=11.9.9

# Telegram
python-telegram-bot>=21.0

# HTTP & async
httpx>=0.27.0
aiofiles>=24.1.0

# Config
pydantic-settings>=2.5.0

# Utils
python-dotenv>=1.0.0
Pillow>=10.0.0
